Move test_plot_explained_variance_ratio to tests/drawing
