+++ hyperspy/tests/drawing/plot_explained_variance_ratio/test_plot_explained_variance_ratio_10-0.001-number-cardinal.png
