+++ hyperspy/tests/drawing/plot_explained_variance_ratio/test_plot_explained_variance_ratio_50-0-number-cardinal.png
